Vantage error handling
