+++ pylabrobot/liquid_handling/backends/__init__.py
@@ -5,7 +5,6 @@ from .serializing_backend import SerializingBackend, SerializingSavingBackend # 
 from .websocket import WebSocketBackend # simulation relies on websocket backend
 
 from .USBBackend import USBBackend
-from .hamilton.errors import HamiltonError
 from .hamilton.STAR import STAR
 from .hamilton.vantage import Vantage
 from .http import HTTPBackend
